update2 sprint4

+++ sprint4/js/main.js
@@ -1,22 +1,24 @@
 $(document).ready(function() {
     $('form').submit(onSubmit);
+    $('[name="stele"]').stars();
     drawTable(store);
+
 });
 //------------------------SUBMIT FORM-------------------------------------//
 var onSubmit = function(){
     var data = {
         city: $('input[name="city"]').val(),
         visited: $('input[name="visited"]').is(":checked"),
-        stars: 5
+        stars: $('input[name="stele"]').val()
     };
     store.add(data).then(function() {
         drawTable(store);
     });
-    $('#inputCity').val('');
-
+    resetForm();
+    $('form').trigger('reset');
     return false;
 };
-//---------------------------------ADD ROWS-------------------------------------//
+//---------------------------------DRAW TABLE-------------------------------------//
 var drawTable= function(store) {
     $('table').find('tbody').empty();
     store.getAll().then(function(data) {
@@ -28,11 +30,70 @@ var drawTable= function(store) {
     });
 };
 var attachEvents = function(){
-    $('table a.delete').click( function(){
+    $('a.delete').confirm({
+        message: 'Are you sure?',
+        onConfirm: function() {
+            //$('table a.delete').click(function(){
+                var id = $(this).closest('tr').data('id');
+                store.delete(id).then(function() {
+                    drawTable(store);
+                });
+                return false;
+            //});
+        },
+        onReject: function() {
+        }
+    });
+    var editedItem = null;
+
+    $('a.edit').click(function () {
         var id = $(this).closest('tr').data('id');
-        store.delete(id).then(function() {
-                drawTable(store);
+        store.getAll().then(function (data) {
+            $.each(data, function () {
+                if (this.id == id) {
+                    editedItem=this;
+                    $('input[name="city"]').val(this.city);
+                    $('input[name="visited"]').prop( "checked", this.visited );
+                    $('input[name="stele"]').val(this.stars).change();
+                    //console.log(id);
+                }
             });
+            return false;
+        });
+        $("a.cancel").show().click(function(){
+            //$('form').trigger('reset');
+            resetForm();
+            $("a.cancel").hide();
+            $(".update").hide();
+            $(".save").show();
+            //console.log(id)
+        });
+        $(".save").hide();
+       $(".update").show().click(function(){
+           var updateData = {
+               city: $('input[name="city"]').val(),
+               visited: $('input[name="visited"]').is(":checked"),
+               stars: $('input[name="stele"]').val(),
+               id:id
+           };
+           store.update(id,updateData).then(function() {
+              drawTable(store);
+           });
+           $(".update").hide();
+           $("a.cancel").hide();
+           $(".save").show();
+           resetForm();
+           //$('form').trigger('reset');
+           //console.log(id)
+           return false;
+       });
         return false;
     });
+      return false;
+};
+
+var resetForm = function (){
+    $('#inputCity').val('');
+    $('input[name="stele"]').val('').change();
+    $('input[name="visited"]').prop('checked', false);
 };
